requirements.txt
# API
fastapi==0.121.1
uvicorn==0.38.0
requests==2.32.5

# Google Cloud
google-cloud-bigquery==3.38.0
google-cloud-core==2.5.0

# ML
pandas==2.3.3
numpy==2.3.4
scikit-learn==1.7.2
implicit==0.7.2
scipy==1.16.3

#.env
# dotenv=0.9.9  # Not sure if needed. the one below is needed to read env
          # variables in jupyter notebooks
python-dotenv==1.2.1

# Needed for importing data from bigquery
db-dtypes==1.4.4
python-multipart==0.0.20

# OCR
easyocr==1.7.2
Pillow==11.0.0
opencv-python==4.10.0.84
google-generativeai==0.8.3
fuzzywuzzy==0.18.0
python-Levenshtein==0.21.1
